requirements.txt
Django>=1.8.2
selenium>=2.46.0
django-autofixture>=0.10.0
